HospitalConsultancySystem
This was very much so a logic based project, code could be tidier but it isn't bad at all working with the time constraints I had at the time. The code is designed around the use cases for a hospital to keep track of patients and consultants and the visits to the hospital.